Fix flaky EndToEndTestWithTls.NoUndecryptablePackets
This test was failing on rare occasions because due to reordering the server would end up receiving a retransmission of the CHLO after it completed the handshake and discarded initial keys. Since QuicConnectionStats::undecryptable_packets_received had been added solely for the purpose of this test and isn't used elsewhere, we replaced it with undecryptable_packets_received_before_handshake_complete which is what we want for this test. Instead of testing for the absence of undecryptable packets at all, we now only test for them before the handshake. This still covers the regression that this test was created for.

gfe-relnote: change statistic that is only used by tests
PiperOrigin-RevId: 269444711
Change-Id: Ie846162b6bf7b65c8c3f7196b9cd58e29d45d892