Allow long header lengths in versions without IETF frames
This is safe without an explicit flag because QuicFramer::WriteIetfLongHeaderLength returns true when the version does not support long header lengths, which is the case for all versions other than v99.

http://google3/third_party/quic/core/quic_framer.cc?l=826&rcl=268972396

The goal of this change is to allow creating v49 which supports long header lengths but not IETF frames.

gfe-relnote: refactor without behavior change, protected by disabled v99 flag
PiperOrigin-RevId: 269738816
Change-Id: I5b69f7a0393e81856a7dc1efcd4a5baa1ebb8dab