[cheri] Change frames to be referenced rather than inlined.
On CHERI architectues the greater wdith and stronger alignment of
capabilities results in all the QUIC frames being greater in size than
a referenced (non-inline) frame. Changing the frames to be referenced
rather than inline is fairly disruptive, as it requires widespread
changes to the access of frame properties and the construction of the
frames. A less disruptive change might be to change the size limit for
inline frames, however the implications of that change for example on
performance aren't clear at present.